refactored downlaod methods to client class. progress bar for aspera. fixed script name. refactored pattern check.
