+++ command-line-downloader/.gitlab-ci.yml
@@ -1,7 +1,7 @@
 image: openjdk:8
 
 variables:
-  APP_VERSION: "1.0.1"
+  APP_VERSION: "0.0.2"
   APP_NAME: 'ena-file-downloader'
   FTP_DIR: "/nfs/ftp/pub/databases/ena/tools/"
