+++ command-line-downloader/src/main/java/uk/ac/ebi/ena/app/utils/CommonUtils.java
@@ -1,8 +1,10 @@
 package uk.ac.ebi.ena.app.utils;
 
 
+import lombok.SneakyThrows;
 import me.tongfei.progressbar.ProgressBarBuilder;
 import me.tongfei.progressbar.ProgressBarStyle;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.SystemUtils;
 import uk.ac.ebi.ena.app.constants.Constants;
 import uk.ac.ebi.ena.app.menu.enums.AccessionTypeEnum;
@@ -39,42 +41,28 @@ public class CommonUtils {
                 .setInitialMax(size);
     }
 
-    public static Map<String, List<String>> getAccessionDetails(List<String> accessions) {
+    @SneakyThrows
+    public static Map<String, List<String>> processAccessions(List<String> accessions) {
         String baseAccession = accessions.get(0);
-        String regex = "";
-        String accessionField = "";
         Map<String, List<String>> accessionDetailsMap = null;
-        if (Pattern.matches(AccessionTypeEnum.EXPERIMENT.getPattern(), baseAccession)) {
-            regex = AccessionTypeEnum.EXPERIMENT.getPattern();
-            accessionField = AccessionTypeEnum.EXPERIMENT.getAccessionField();
-        } else if (Pattern.matches(AccessionTypeEnum.SAMPLE.getPattern(), baseAccession)) {
-            regex = AccessionTypeEnum.SAMPLE.getPattern();
-            accessionField = AccessionTypeEnum.SAMPLE.getAccessionField();
-        } else if (Pattern.matches(AccessionTypeEnum.STUDY.getPattern(), baseAccession)) {
-            regex = AccessionTypeEnum.STUDY.getPattern();
-            accessionField = AccessionTypeEnum.STUDY.getAccessionField();
-        } else if (Pattern.matches(AccessionTypeEnum.RUN.getPattern(), baseAccession)) {
-            regex = AccessionTypeEnum.RUN.getPattern();
-            accessionField = AccessionTypeEnum.RUN.getAccessionField();
-        } else if (Pattern.matches(AccessionTypeEnum.ANALYSIS.getPattern(), baseAccession)) {
-            regex = AccessionTypeEnum.ANALYSIS.getPattern();
-            accessionField = AccessionTypeEnum.ANALYSIS.getAccessionField();
+        AccessionTypeEnum type = AccessionTypeEnum.getAccessionTypeByPattern(baseAccession);
+        if (type != null) {
+            boolean isValid = accessions.stream().allMatch(acc -> Pattern.matches(type.getPattern(), acc));
+            if (isValid) {
+                accessionDetailsMap = getAccessionDetailsMap(type.getAccessionField(), accessions);
+            }
+            return accessionDetailsMap;
+        } else {
+            System.out.println("Unsupported accession type provided:" + baseAccession + " Supported types are "
+                    + StringUtils.join(AccessionTypeEnum.values(), ", "));
+            throw new Exception("Unsupported accession type provided:" + baseAccession);
         }
-        String finalRegex = regex;
-
-        boolean isValid = accessions.stream().allMatch(acc -> Pattern.matches(finalRegex, acc));
-        if (isValid) {
-            accessionDetailsMap = getAccessionDetailsMap(accessionField, accessions);
-        }
-
-        return accessionDetailsMap;
     }
 
     public static Map<String, List<String>> getAccessionDetailsMap(String accessionField, List<String> accessions) {
         Map<String, List<String>> accessionDetailsMap = new HashMap<>();
         accessionDetailsMap.put(ACCESSION_FIELD, Collections.singletonList(accessionField));
         accessionDetailsMap.put(ACCESSION_LIST, accessions);
-
         return accessionDetailsMap;
     }
